+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.8.0 - 2022-08-19
+### Feature
+- 하이브리드/항공 지도 유형 설정 API 추가
+  - `InaviMap#MapType`, `InaviMap#setMapType`, `InaviMap#getMapType`, `InvMapOptions#setMapType`, `InvMapOptions#getMapType`
+- 좌표 변환 API 추가
+  - `Katec`, `Utmk`, `Tm`, `Grs80`
+
 ## 0.7.1 - 2022-03-30
 ### Feature
 - 로고의 위치와 마진을 설정하는 API 추가